Refactor to remove spring boot from commons
